refactor(shared): restore public API boundaries (#913)
