+++ src/entities/card/api/subscription.spec.tsx
@@ -23,7 +23,7 @@ vi.mock("firebase/firestore", async (importOriginal) => {
     where: mocks.where,
   };
 });
-vi.mock("@/shared/api/firebase", () => ({ db: "db" }));
+vi.mock("@/shared/firebase", () => ({ db: "db" }));
 
 import { subscribeCards } from "./firestore";
