+++ src/entities/deck/api/firestore.ts
@@ -11,8 +11,8 @@ import type {
 import { collection, deleteDoc, doc, getDocs, onSnapshot, query, setDoc, updateDoc, where } from "firebase/firestore";
 import { z } from "zod";
 
-import { db } from "@/shared/api/firebase";
-import { getTimestamp, omitUndefined, parseFirestoreDocument } from "@/shared/api/firestoreDocument";
+import { db } from "@/shared/firebase";
+import { getTimestamp, omitUndefined, parseFirestoreDocument } from "@/shared/api";
 import { createDeckSchema, deleteDeckSchema, editDeckSchema } from "../model/schema";
 import { replaceDecks } from "../model/store";
