+++ src/entities/study-progress/api/firestore.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from "vitest";
 
-vi.mock("@/shared/api/firebase", () => ({ db: {} }));
+vi.mock("@/shared/firebase", () => ({ db: {} }));
 
 import { editStudyProgress } from "./firestore";
