+++ src/entities/study-progress/api/firestore.ts
@@ -2,8 +2,8 @@ import type { EditStudyProgressInput } from "../model/types";
 
 import { doc, updateDoc } from "firebase/firestore";
 
-import { db } from "@/shared/api/firebase";
-import { getTimestamp, omitUndefined } from "@/shared/api/firestoreDocument";
+import { db } from "@/shared/firebase";
+import { getTimestamp, omitUndefined } from "@/shared/api";
 import { editStudyProgressSchema } from "../model/schema";
 
 export const editStudyProgress = async (uid: string, progress: EditStudyProgressInput["progress"]): Promise<void> => {
