+++ src/features/card-list/ui/CardListView.spec.tsx
@@ -13,7 +13,7 @@ import { describe, expect, it, vi } from "vitest";
 
 import { createCard } from "@/test/factories";
 
-vi.mock("@/shared/api/firebase", () => ({ auth: {} }));
+vi.mock("@/shared/firebase", () => ({ auth: {} }));
 
 import { CardListView } from "./CardListView";
