+++ src/features/study/hooks/useStudyActions.spec.tsx
@@ -11,7 +11,7 @@ import type { Preferences } from "@/entities/preferences";
 import { act, renderHook } from "@testing-library/react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-vi.mock("@/shared/api/firebase", () => ({ auth: {}, db: {} }));
+vi.mock("@/shared/firebase", () => ({ auth: {}, db: {} }));
 
 import { useStudyActions } from "./useStudyActions";
 import { studyStore } from "../state/studyStoreInstance";
