+++ src/features/study/hooks/useStudyCards.spec.tsx
@@ -1,7 +1,7 @@
 import { act, renderHook } from "@testing-library/react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-vi.mock("@/shared/api/firebase", () => ({ auth: {}, db: {} }));
+vi.mock("@/shared/firebase", () => ({ auth: {}, db: {} }));
 
 import { createCard, createPreferences, createDeck } from "@/test/factories";
