+++ src/pages/card-view/ui/CardViewPage.spec.tsx
@@ -18,7 +18,7 @@ const mocks = vi.hoisted(() => ({
   setDarkMode: vi.fn(),
 }));
 
-vi.mock("@/shared/api/firebase", () => ({ auth: {} }));
+vi.mock("@/shared/firebase", () => ({ auth: {} }));
 vi.mock("@/entities/preferences", () => ({
   usePreferences: () => mocks.preferences,
   setDarkMode: mocks.setDarkMode,
