+++ src/pages/card-view/ui/CardViewView.spec.tsx
@@ -8,7 +8,7 @@ import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom/vitest";
 import { describe, expect, it, vi } from "vitest";
 
-vi.mock("@/shared/api/firebase", () => ({ auth: {} }));
+vi.mock("@/shared/firebase", () => ({ auth: {} }));
 
 import { CardViewView } from "./CardViewView";
