+++ src/pages/deck-form/ui/DeckFormPage.spec.tsx
@@ -13,7 +13,7 @@ const mocks = vi.hoisted(() => ({
   navigate: vi.fn(),
 }));
 
-vi.mock("@/shared/api/firebase", () => ({ auth: {}, db: {} }));
+vi.mock("@/shared/firebase", () => ({ auth: {}, db: {} }));
 
 vi.mock("@/entities/deck", async (importOriginal) => {
   const actual = await importOriginal<typeof import("@/entities/deck")>();
