+++ src/pages/deck-list/ui/DeckListPage.spec.tsx
@@ -81,7 +81,7 @@ vi.mock("@/features/deck-list", () => ({
   },
 }));
 vi.mock("react-router-dom", () => ({ useNavigate: () => mocks.navigate }));
-vi.mock("@/shared/api/firebase", () => ({ auth: {}, db: {} }));
+vi.mock("@/shared/firebase", () => ({ auth: {}, db: {} }));
 
 import { DeckListPage } from "./DeckListPage";
