+++ src/pages/deck-swiper/ui/DeckSwiperPage.spec.tsx
@@ -45,7 +45,7 @@ const mocks = vi.hoisted(() => ({
   setDarkMode: vi.fn(),
 }));
 
-vi.mock("@/shared/api/firebase", () => ({ auth: {} }));
+vi.mock("@/shared/firebase", () => ({ auth: {} }));
 
 vi.mock("@/entities/preferences", () => ({
   usePreferences: () => {
