+++ src/shared/firebase/index.ts
