+++ test/integration/firestore/card.spec.ts
@@ -18,17 +18,17 @@ import { expect, it, describe, vi, beforeEach, type Mock } from "vitest";
 import { collection, deleteDoc, getDocs, getFirestore, doc, getDoc, query, where } from "firebase/firestore";
 import { upsertImportedCards } from "@/features/deck-import/api/upsertImportedCards";
 import { editStudyProgress } from "@/entities/study-progress";
-import { getTimestamp } from "@/shared/api/firestoreDocument";
+import { getTimestamp } from "@/shared/api";
 import * as UUID from "uuid";
 import { createCard, createDeck } from "@/test/factories";
 
 const uuid = UUID.v4;
 
-vi.mock("@/shared/api/firestoreDocument", async (importOriginal) => ({
-  ...(await importOriginal<typeof import("@/shared/api/firestoreDocument")>()),
+vi.mock("@/shared/api", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("@/shared/api")>()),
   getTimestamp: vi.fn(),
 }));
-vi.mock("@/shared/api/firebase", async () => ({
+vi.mock("@/shared/firebase", async () => ({
   db: (await import("@/test/initializeTestFirestore")).testDb,
 }));
