+++ test/integration/firestore/deck.spec.ts
@@ -9,17 +9,17 @@ import "@/test/initializeTestFirestore";
 import { expect, it, describe, vi, beforeEach, type Mock } from "vitest";
 import { doc, getDoc, getFirestore } from "firebase/firestore";
 import { createCard as createCardCommand } from "@/entities/card";
-import { getTimestamp } from "@/shared/api/firestoreDocument";
+import { getTimestamp } from "@/shared/api";
 import * as UUID from "uuid";
 import { createCard, createDeck as createDeckFixture } from "@/test/factories";
 
 const uuid = UUID.v4;
 
-vi.mock("@/shared/api/firestoreDocument", async (importOriginal) => ({
-  ...(await importOriginal<typeof import("@/shared/api/firestoreDocument")>()),
+vi.mock("@/shared/api", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("@/shared/api")>()),
   getTimestamp: vi.fn(),
 }));
-vi.mock("@/shared/api/firebase", async () => ({
+vi.mock("@/shared/firebase", async () => ({
   db: (await import("@/test/initializeTestFirestore")).testDb,
 }));
